rename and re-organise filter apimonitoring
